+++ src/Text/index.js
